[Feat] 아이템 착용하기를 선택할 경우 착용한 뒤 MyPageVC로 이동
