+++ BJJ_iOS/BJJ_iOS/Source/Network/API/MyPage/MyPage/MyPageAPI.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import Alamofire
 
 final class MyPageAPI {
     static func fetchMyPageInfo(completion: @escaping (Result<MyPage, Error>) -> Void) {
@@ -15,17 +14,6 @@ final class MyPageAPI {
             method: .get,
             data: nil,
             model: MyPage.self,
-            completion: completion
-        )
-    }
-    
-    static func patchItem(itemType: String, itemID: Int, completion: @escaping (Result<Empty, Error>) -> Void) {
-        networkRequest(
-            urlStr: MyPageAddress.patchItem(itemID).url,
-            method: .patch,
-            data: nil,
-            model: Empty.self,
-            query: ["itemType": itemType],
             completion: completion
         )
     }
