+++ BJJ_iOS/BJJ_iOS/Source/Network/API/MyPage/MyPage/MyPageAddress.swift
@@ -9,14 +9,11 @@ import Foundation
 
 enum MyPageAddress {
     case fetchMyPageInfo
-    case patchItem(Int)
     
     var url: String {
         switch self {
         case .fetchMyPageInfo:
             return "items/my"
-        case .patchItem(let itemID):
-            return "items/\(itemID)"
         }
     }
 }
