+++ BJJ_iOS/BJJ_iOS/Source/Presentation/Tab/MyPage/GachaResult/Model/GachaResultSection.swift
@@ -1,0 +1,13 @@
+//
+//  GachaResultSection.swift
+//  BJJ_iOS
+//
+//  Created by HyoTaek on 4/14/25.
+//
+
+import Foundation
+
+struct GachaResultSection: Hashable {
+    let itemID: Int
+    let itemType: String
+}
